+++ resources/bootstrap/profiles/core/docker/compose-project-name.sh
@@ -2,13 +2,14 @@
 
 set -eu -o pipefail
 
+SCRIPT_DIR="$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")"
+
 # docker-compose versions prior to 1.21.0 stripped dashes and
 # underscores from project names. Later versions don't. As we want to
 # build the same project name docker-compose would, we need to use a
 # slightly different sed expression depending on the docker-compose
 # version we are using.
-compose_version="$(docker-compose --version |
-    sed 's/docker-compose version \([0-9][0-9]*\.[0-9][0-9]*\.[0-9][0-9]*\).*/\1/g')"
+compose_version="$("${SCRIPT_DIR}/docker-compose.sh" version --short)"
 if [[ "${compose_version}" < "1.21.0" ]]; then
     SED_EXPR='s/[^a-z0-9]//g'
 else
@@ -22,7 +23,7 @@ fi
 # the parent directory for the script. From there we take the simple
 # name (base name) to get the name of the parent directory, where all
 # the docker-compose files are.
-DIR_NAME="$(basename "$(dirname "$(dirname "$(readlink -f "$0")")")")"
+DIR_NAME="$(basename "$(dirname "$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")")")"
 COMPOSE_PROJECT="$(echo -n "${DIR_NAME}" | tr '[:upper:]' '[:lower:]' | sed "${SED_EXPR}")"
 
 # Write project name to stdout, so the calling script can capture it.