+++ resources/bootstrap/profiles/core/start-dev.sh
@@ -12,7 +12,7 @@ export COMPOSE_FILE="{{project.docker-compose.to-develop}}"
 compose_project="$(docker/compose-project-name.sh)"
 
 # Stop any containers still running. Don't wait for them to finish :-)
-docker-compose down --timeout 0
+docker/docker-compose.sh down --timeout 0
 
 # Clean up any left overs from previous runs, to make sure we start
 # with a clean environment (stale Docker containers, stale Clojure
@@ -40,14 +40,14 @@ fi
 # is going to be created owned by root (because it is a bind mount
 # point). And that is going to prevent downloading any dependencies!
 mkdir -p ~/.m2/
-docker-compose run --no-deps --rm app lein clean
+docker/docker-compose.sh run --no-deps --rm app lein clean
 
 # Make sure we are not trying to use any environment vars in
 # docker-compose.yml that are not set.
-./docker/docker-env-vars.sh
+docker/docker-env-vars.sh
 
 # Finally launch the application itself
-docker-compose up --build --detach --force-recreate --renew-anon-volumes
+docker/docker-compose.sh up --build --detach --force-recreate --renew-anon-volumes
 
 # And show the logs
-docker-compose logs --follow --timestamps
+docker/docker-compose.sh logs --follow --timestamps