+++ resources/bootstrap/profiles/on-premises/on-premises-files/usr/local/hop/{{project.name}}/bin/start-app.sh
@@ -3,11 +3,11 @@
 set -eu -o pipefail
 
 # Stop any containers still running. Don't wait for them to finish :-)
-docker-compose down -t0
+docker/docker-compose.sh down -t0
 
 # Make sure to check that we are not trying to use any environment
 # vars in docker-compose.yml that are not set.
-docker-compose up -d --force-recreate --renew-anon-volumes
+docker/docker-compose.sh up -d --force-recreate --renew-anon-volumes
 
 # And show the logs
-docker-compose logs -ft --no-color
+docker/docker-compose.sh logs -ft --no-color