+++ resources/bootstrap/profiles/on-premises/on-premises-files/usr/local/hop/{{project.name}}/bin/stop-app.sh
@@ -3,4 +3,4 @@
 set -eu -o pipefail
 
 ## Stop any containers still running. Don't wait for them to finish :-)
-docker-compose down -t0
+docker/docker-compose.sh down -t0